requirements.txt
faiss-cpu==1.8.0.post1
numpy==1.26.4
openai==1.51.2
pydantic==2.9.2
pypdf==5.0.1
python-dotenv==1.0.1
sentence-transformers==3.1.1
